Conflicts:
	src/particle_creator/head/pc_file.h
	src/particle_creator/source/pc_file.cpp

merging wth master?

+++ src/basic/head/data_format.h
@@ -165,13 +165,13 @@ public:
     /*!
      * Index of the attribute.
      */
-    int index;
+    size_t index;
 
     /*!
      * Start of the attribute's data from the beginning of the overall data
      * stored in a \a Data.
      */
-    ptrdiff_t offset;
+    size_t offset;
 
     /*!
      * Type of data.
